Cambios

+++ src/main/java/ManyWorker/service/CategoriaService.java
@@ -2,6 +2,8 @@ package ManyWorker.service;
 
 import ManyWorker.entity.Categoria;
 import ManyWorker.repository.CategoriaRepository;
+import ManyWorker.repository.TareaReparacionRepository;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +15,9 @@ public class CategoriaService {
 
     @Autowired
     private CategoriaRepository categoriaRepository;
+    
+    @Autowired
+    private TareaReparacionRepository tareaReparacionRepository;
 
     // Crear una nueva Categoria
     public Categoria crearCategoria(Categoria categoria) {
@@ -36,10 +41,16 @@ public class CategoriaService {
     public List<Categoria> listarCategorias() {
         return categoriaRepository.findAll();
     }
-
-    // Eliminar una categoria
+    
+ // Eliminar una categoria
     public void eliminarCategoria(int categoriaId) {
-    	categoriaRepository.deleteById(categoriaId);
+        // Verificar si la categoría tiene tareas de reparación asociadas
+        boolean tieneTareasAsociadas = tareaReparacionRepository.existsById(categoriaId);
+        if (tieneTareasAsociadas) {
+            throw new RuntimeException("No se puede eliminar una categoría con tareas de reparación asociadas.");
+        }
+
+        categoriaRepository.deleteById(categoriaId);
     }
 
     // Obtener una Categoria por ID
